+++ getmac/__main__.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .getmac import get_mac_address, __version__
 import argparse
+import sys
+
+import getmac
 
 
 def main():
@@ -15,8 +17,8 @@ def main():
                     'or network interfaces using Python.',
         prog='get-mac', add_help=True)
     parser.add_argument('--version', action='version',
-                        version='get-mac %s' % __version__)
-    parser.add_argument('-d', '--no-network-requests', action='store_true',
+                        version='get-mac %s' % getmac.getmac.__version__)
+    parser.add_argument('--no-network-requests', action='store_true',
                         help='Disable refreshing of the ARP '
                              'table by making a network request (ping)')
     group = parser.add_mutually_exclusive_group(required=True)
@@ -28,12 +30,18 @@ def main():
                        help='IPv6 address of a remote host')
     group.add_argument('--hostname', type=str, default=None,
                        help='Hostname of a remote host')
+    parser.add_argument('--debug', action='store_true',
+                        help='For debugging failures')
     args = parser.parse_args()
 
-    mac = get_mac_address(interface=args.interface, ip=args.ip,
-                          ip6=args.ip6, hostname=args.hostname,
-                          network_request=args.no_network_requests)
-    print(mac)
+    if args.debug:
+        getmac.getmac.DEBUG = True
+
+    mac = getmac.get_mac_address(interface=args.interface, ip=args.ip,
+                                 ip6=args.ip6, hostname=args.hostname,
+                                 network_request=args.no_network_requests)
+    print("" if mac is None else mac)
+    sys.exit(0)
 
 
 if __name__ == '__main__':